chore: remove unshipped server-push subsystem; notifications are local-only
Product decision: reminders stay on-device (UNUserNotificationCenter);
there will be no server-sent push, so the never-wired device-token
groundwork comes out instead of waiting for a feature that isn't coming:

- shared/crypto.ts -> shared/redact.ts: keep redactSensitiveFields()
  (general log hygiene), drop pepper init + token hash/verify/validate
- types.ts: drop DeviceRegistrationInput
- firebase.ts: drop COLLECTIONS.devices
- env.ts: drop NOTIFICATION_TOKEN_PEPPER_SECRET_NAME
- firestore.rules: drop the devices deny block (catch-all covers it) and
  the stale token bullets in the header comment
- delete functions/NOTIFICATION_TOKEN_SECURITY.md
- CLAUDE.md: correct endpoint/collection lists, document the
  local-only notifications decision

Verified: tsc exit 0; rules load clean in the Firestore emulator.
