+++ functions/src/shared/env.ts
@@ -11,4 +11,3 @@ export const GCP_PROJECT = process.env.GCP_PROJECT || process.env.GCLOUD_PROJECT
 // SECURITY: Only export secret NAME, never the actual key
 // The actual key must be retrieved via getSecretString() from secrets.ts
 export const OPENAI_API_KEY_SECRET_NAME = process.env.OPENAI_API_KEY_SECRET_NAME;
-export const NOTIFICATION_TOKEN_PEPPER_SECRET_NAME = process.env.NOTIFICATION_TOKEN_PEPPER_SECRET_NAME;
